# shared/openapi/inventory.yaml
openapi: 3.1.0
info:
  title: Inventory Service API
  description: Multi-tenant inventory management and warehouse operations service
  contact:
    name: Anthill Team
    email: team@example.com
  license:
    name: MIT
  version: 0.1.0
servers:
- url: http://localhost:8001
  description: Local development server
paths:
...
  /api/v1/inventory/valuation/{product_id}:
    get:
      tags:
      - valuation
      summary: GET /api/v1/inventory/valuation/{product_id} - Get current valuation for a product
      description: |-
        Returns the current inventory valuation for a specific product.
        The valuation includes current quantity, value, and cost based on the
        product's valuation method (FIFO, AVCO, or Standard).

        # Authentication
        Requires authenticated user with appropriate tenant access

        # Path Parameters
        * `product_id` - Product UUID

        # Returns
        * `200` - Current valuation data
        * `404` - Product or valuation not found
        * `401` - Authentication required
        * `403` - Insufficient permissions
      operationId: get_valuation
      parameters:
      - name: product_id
        in: path
        description: Product ID
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Current valuation data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValuationDto'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Product or valuation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_auth: []
  /api/v1/inventory/valuation/{product_id}/adjust:
    post:
      tags:
      - valuation
      summary: POST /api/v1/inventory/valuation/{product_id}/adjust - Adjust inventory cost
      description: |-
        Performs a cost adjustment to the inventory valuation.
        This can be used for write-offs, revaluations, or corrections.

        # Authentication
        Requires authenticated user with appropriate tenant access

        # Path Parameters
        * `product_id` - Product UUID

        # Request Body
        ```json
        {
          "adjustment_amount": -5000,  // Adjustment in cents (can be negative)
          "reason": "Damaged goods write-off"
        }
        ```

        # Returns
        * `200` - Updated valuation data
        * `400` - Invalid adjustment amount
        * `404` - Product not found
        * `401` - Authentication required
        * `403` - Insufficient permissions
      operationId: adjust_cost
      parameters:
      - name: product_id
        in: path
        description: Product ID
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CostAdjustmentPayload'
        required: true
      responses:
        '200':
          description: Updated valuation data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValuationDto'
        '400':
          description: Invalid adjustment amount
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_auth: []
  /api/v1/inventory/valuation/{product_id}/history:
    get:
      tags:
      - valuation
      summary: GET /api/v1/inventory/valuation/{product_id}/history - Get valuation history
      description: |-
        Returns the historical changes to a product's valuation.
        Useful for auditing and financial reporting.

        # Authentication
        Requires authenticated user with appropriate tenant access

        # Path Parameters
        * `product_id` - Product UUID

        # Query Parameters
        * `limit` - Maximum number of records (default: 50, max: 100)
        * `offset` - Number of records to skip (default: 0)

        # Returns
        * `200` - Historical valuation records with pagination
        * `404` - Product not found
        * `401` - Authentication required
        * `403` - Insufficient permissions
      operationId: get_valuation_history
      parameters:
      - name: product_id
        in: path
        description: Product ID
        required: true
        schema:
          type: string
          format: uuid
      - name: limit
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int64
      - name: offset
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int64
      responses:
        '200':
          description: Historical valuation records with pagination
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValuationHistoryResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_auth: []
  /api/v1/inventory/valuation/{product_id}/layers:
    get:
      tags:
      - valuation
      summary: GET /api/v1/inventory/valuation/{product_id}/layers - Get valuation layers for FIFO
      description: |-
        Returns the active cost layers for products using FIFO valuation.
        Layers represent different cost levels for remaining inventory.

        # Authentication
        Requires authenticated user with appropriate tenant access

        # Path Parameters
        * `product_id` - Product UUID

        # Returns
        * `200` - List of active cost layers
        * `404` - Product not found or doesn't use FIFO
        * `401` - Authentication required
        * `403` - Insufficient permissions
      operationId: get_valuation_layers
      parameters:
      - name: product_id
        in: path
        description: Product ID
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: List of active cost layers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValuationLayersResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Product not found or doesn't use FIFO
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_auth: []
  /api/v1/inventory/valuation/{product_id}/method:
    put:
      tags:
      - valuation
      summary: PUT /api/v1/inventory/valuation/{product_id}/method - Set valuation method for a product
      description: |-
        Changes the valuation method for a product. This affects how inventory
        costs are calculated and tracked.

        # Authentication
        Requires authenticated user with appropriate tenant access

        # Path Parameters
        * `product_id` - Product UUID

        # Request Body
        ```json
        {
          "valuation_method": "fifo" | "avco" | "standard"
        }
        ```

        # Returns
        * `200` - Updated valuation data
        * `400` - Invalid valuation method
        * `404` - Product not found
        * `401` - Authentication required
        * `403` - Insufficient permissions
      operationId: set_valuation_method
      parameters:
      - name: product_id
        in: path
        description: Product ID
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetValuationMethodPayload'
        required: true
      responses:
        '200':
          description: Updated valuation data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValuationDto'
        '400':
          description: Invalid valuation method
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_auth: []
  /api/v1/inventory/valuation/{product_id}/revalue:
    post:
      tags:
      - valuation
      summary: POST /api/v1/inventory/valuation/{product_id}/revalue - Revalue inventory
      description: |-
        Revalues the entire inventory at a new cost basis.
        This changes the cost of existing inventory without affecting quantity.

        # Authentication
        Requires authenticated user with appropriate tenant access

        # Path Parameters
        * `product_id` - Product UUID

        # Request Body
        ```json
        {
          "new_unit_cost": 12000,  // New cost per unit in cents
          "reason": "Market price adjustment"
        }
        ```

        # Returns
        * `200` - Updated valuation data
        * `400` - Invalid cost
        * `404` - Product not found
        * `401` - Authentication required
        * `403` - Insufficient permissions
      operationId: revalue_inventory
      parameters:
      - name: product_id
        in: path
        description: Product ID
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevaluationPayload'
        required: true
      responses:
        '200':
          description: Updated valuation data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValuationDto'
        '400':
          description: Invalid cost
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_auth: []
  /api/v1/inventory/valuation/{product_id}/standard-cost:
    put:
      tags:
      - valuation
      summary: PUT /api/v1/inventory/valuation/{product_id}/standard-cost - Set standard cost for a product
      description: |-
        Sets the standard cost for products using Standard costing method.
        Only applicable when the product uses standard valuation.

        # Authentication
        Requires authenticated user with appropriate tenant access

        # Path Parameters
        * `product_id` - Product UUID

        # Request Body
        ```json
        {
          "standard_cost": 15000  // Cost in cents
        }
        ```

        # Returns
        * `200` - Updated valuation data
        * `400` - Invalid cost or product doesn't use standard costing
        * `404` - Product not found
        * `401` - Authentication required
        * `403` - Insufficient permissions
      operationId: set_standard_cost
      parameters:
      - name: product_id
        in: path
        description: Product ID
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetStandardCostPayload'
        required: true
      responses:
        '200':
          description: Updated valuation data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValuationDto'
        '400':
          description: Invalid cost or product doesn't use standard costing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer_auth: []
...
components:
  schemas:
...
    CostAdjustmentPayload:
      type: object
      required:
      - adjustment_amount
      - reason
      properties:
        adjustment_amount:
          type: integer
          format: int64
        reason:
          type: string
...
    ErrorResponse:
      type: object
      description: Error response for OpenAPI documentation
      required:
      - error
      - code
      properties:
        code:
          type: string
          description: Error code
        error:
          type: string
          description: Error message
...
    RevaluationPayload:
      type: object
      required:
      - new_unit_cost
      - reason
      properties:
        new_unit_cost:
          type: integer
          format: int64
        reason:
          type: string
...
    SetStandardCostPayload:
      type: object
      required:
      - standard_cost
      properties:
        standard_cost:
          type: integer
          format: int64
    SetValuationMethodPayload:
      type: object
      required:
      - valuation_method
      properties:
        valuation_method:
          $ref: '#/components/schemas/ValuationMethod'
...
    ValuationDto:
      type: object
      description: Current valuation information for a product
      required:
      - valuation_id
      - tenant_id
      - product_id
      - valuation_method
      - total_quantity
      - total_value
      - last_updated
      properties:
        current_unit_cost:
          type:
          - integer
          - 'null'
          format: int64
        last_updated:
          type: string
          format: date-time
        product_id:
          type: string
          format: uuid
        standard_cost:
          type:
          - integer
          - 'null'
          format: int64
        tenant_id:
          type: string
          format: uuid
        total_quantity:
          type: integer
          format: int64
        total_value:
          type: integer
          format: int64
        valuation_id:
          type: string
          format: uuid
        valuation_method:
          $ref: '#/components/schemas/ValuationMethod'
    ValuationHistoryDto:
      type: object
      description: Historical valuation record
      required:
      - history_id
      - valuation_id
      - tenant_id
      - product_id
      - valuation_method
      - total_quantity
      - total_value
      - changed_at
      properties:
        change_reason:
          type:
          - string
          - 'null'
        changed_at:
          type: string
          format: date-time
        history_id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        standard_cost:
          type:
          - integer
          - 'null'
          format: int64
        tenant_id:
          type: string
          format: uuid
        total_quantity:
          type: integer
          format: int64
        total_value:
          type: integer
          format: int64
        unit_cost:
          type:
          - integer
          - 'null'
          format: int64
        valuation_id:
          type: string
          format: uuid
        valuation_method:
          $ref: '#/components/schemas/ValuationMethod'
    ValuationHistoryResponse:
      type: object
      description: Response for valuation history
      required:
      - history
      - total_count
      properties:
        history:
          type: array
          items:
            $ref: '#/components/schemas/ValuationHistoryDto'
        total_count:
          type: integer
          format: int64
    ValuationLayerDto:
      type: object
      description: Cost layer for FIFO valuation
      required:
      - layer_id
      - tenant_id
      - product_id
      - quantity
      - unit_cost
      - total_value
      - created_at
      properties:
        created_at:
          type: string
          format: date-time
        layer_id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        quantity:
          type: integer
          format: int64
        tenant_id:
          type: string
          format: uuid
        total_value:
          type: integer
          format: int64
        unit_cost:
          type: integer
          format: int64
    ValuationLayersResponse:
      type: object
      description: Response for valuation layers
      required:
      - layers
      properties:
        layers:
          type: array
          items:
            $ref: '#/components/schemas/ValuationLayerDto'
    ValuationMethod:
      type: string
      description: Supported inventory valuation methods
      enum:
      - fifo
      - avco
      - standard
...